Commented out some unused parameters
closes #513

+++ solvers/thermal/dynamic/iterative_matrix.h
@@ -63,7 +63,7 @@ struct SparseBandMatrix {
     void clear() {
         std::fill_n(data, LDA*size, 0.);
     }
-    
+
     /**
      * Multiply matrix by vector
      * \param vector vector to multiply
@@ -194,7 +194,7 @@ int solveDCG(Matrix& matrix, const Preconditioner& msolve, double* x, double* b,
     // Calculate r = b - Ax and initial error.
     matrix.multiply(x, r.get());
 
-    for (int j = 0; j < n; ++j) r[j] = b[j] - r[j];
+    for (size_t j = 0; j < n; ++j) r[j] = b[j] - r[j];
     err = ddot(n, r.get(), 1, r.get(), 1) / bnorm2;
     if (err < eps2) {
         return 0;
@@ -218,7 +218,7 @@ int solveDCG(Matrix& matrix, const Preconditioner& msolve, double* x, double* b,
             bkden = bknum;
 
             // Calculate p = z + bk*p
-            for (int j = 0; j < n; ++j)
+            for (size_t j = 0; j < n; ++j)
                 p[j] = fma(bk, p[j], z[j]);
         }
         // Calculate z = Ap, akden = (p,Ap) and ak.
